+++ app/src/main/java/com/itverse/futuris/data/repositories/ComposantRepository.kt
@@ -7,7 +7,7 @@ import kotlinx.coroutines.flow.Flow
 
 class ComposantRepository (private val composantDao: ComposantDao){
 
-    fun allProjectComposants(projectId: Int): Flow<List<Composant>> {
+    fun allProjectComposants(projectId: Long): Flow<List<Composant>> {
         return composantDao.getAllFromProject(projectId)
     }
